Fix 10550373: Stopping FUL when emergency dialer opens
Face Unlock wasn't being stopped when the emergency dialer opened,
causing FUL to flash briefly after the emergency dialer appeared.
This was most noticeable on Hammerhead devices.

This change allows EmergencyButton to stop FUL by way of
KeyguardUpdateMonitor.  Note that it must stop FUL immediately rather
than going through a handler.  If going through a handler FUL is not
stopped in time to prevent the issue.

Change-Id: I3985969a19e909e2633d7876a3fb2d34c70e2cc1

Former-commit-id: 7fce38021694925295f5d14bfba71f28cba19404