Only show launcher for the bottom activity in a task
When transitioning from activity-over-launcher to task-over-launcher
ensureActivitiesVisibleLocked() was too aggressive in showing the
launcher. If there were any non-fullscreen activities in a task that
sits over the launcher then the launcher would be shown.

This fix adds a test to make sure the launcher will only be shown if
the bottommost activity in such a task is non-fullscreen.

Fixes bug 10840919.

Change-Id: I5dcd63be3fa2865ae38cbb921332937dfa4b5d47

Former-commit-id: f51b5588d795ae6add9f625482d7d3f3ab34fb94