Update date format in panel on locale change.
The logic that updated the locale was only running at times
the DateView should have been updating its display, namely,
when the view was visible on-screen.

The new version trades a small amount of efficiency (the
broadcast receiver is now active at all times, and the date
is re-formatted each TIME_TICK) in exchange for simplicity
and correctness.

Bug: 10725912
Change-Id: I2f0988ed478374ed0eb0c76c3d8caed6c86eb8df

Former-commit-id: b478a71625235f554263324dbf0501f6b6c87520