Fix lockscreen wave animation artifacts
On some display sizes, the wave animation was sometimes
running more than once, starting over in the center and animating
outward... partially.

The problem was that the calculations determining the alpha value
of the dots was returning bogus values when the display area was
large enough, which is why the bug is only on some devices.

This fix simplifies the math and ensures that the wave only animates
once, from start to finish.

Issue #11005936 regression on lockscreen animation for multi-wave widget

Change-Id: Id21a2e4d2271f01c82c4bc6e1f37d78e68b9b6e4

Former-commit-id: 167a32326ea3a3b30c147cd02e3b3151e794e016