Fix screen magnifier scaling due to ScaleGestureDetector changes.
The scale gesture detector has a new behavior to make onScale
callbacks during swipe after a double tap. The screen magnification
is triggered after a trippe tap so if the user trippe taps and
holds to do a temporary magnification and tries to change the
zoom level with another finger, erroneous results are obtained.

The ScaleGestureDetector uses an APIs level check for the new
behavior but the ScreenMagnifier is a platform feature. We now
explicitly ask for the old behavior.

bug:11033376

Change-Id: I0dfb14dd3abcaa34ad1f40447c631b4203797378

Former-commit-id: 1fdbaa100c00e7e93db6f96f8ff9327a0aad968c