Relax conditions for including windows behind dialogs
When a dialog has been minimized to recents the windows behind it
won't be visible. Yet we were requiring them to be visible in order to
be included in the ones being restored. This left the background
windows invisible on resume and showed home behind floating dialogs
instead of the activity that launched the dialogs.

Fixes bug 11067724.

Change-Id: Icadd7ec8fe7c73b52982b6ff5b5d98b8fb8476b0

Former-commit-id: ff174d52bdac8893893f64c772e570842214f663