Fix dock rect calculation when nav is translucent.
Make sure the dock rect in PWM takes the nav bar height
into account when translucent.  This forms the basis of
many other window calculations, and has multiple downstream
effects.

Fixing this exposed the fact that the Keyguard was not allowed
to fall into the LAYOUT_HIDE_NAV codepath (since it's not an app
window) - so add TYPE_KEYGUARD to that path.

Bug:11157555
Change-Id: I6b1fc4da973a246b80ec108a561e06f6d8a6a92b

Former-commit-id: f95b89a0971fdfedec9004924615f08d87a6168c