Fix updating fast scroll state for data set changes
Only looking for old/new count changes is incomplete with the current
state of things. The observer's onChanged method will update this in
some cases, stomping the different values needed to trigger a fast
scroller update. Also update the fast scroller if the data change flag
is set.

Also fix the positioning of the legacy text overlay.

Bug 11188512

Change-Id: I35b3915ce49f8494c6d82f6be6d6df7169deddd7

Former-commit-id: 5db566f16a1b8f36e84a9be00cde40482d48466b