Don't move clock when warping in PagedView
This fixes a bug where the clock would appear to move when flinging
the camera widget while warping.  To accomplish this,  it no longer
scrolls while warping but rather updates the warp view's translationX.
When the finger is released or a fling happens, we either smoothly
animate the view back (without scrolling) or update the scroll position
and let the scroller take over the animation.

Fixes bug 11191094

Change-Id: I33f087cf4fb2ec31db11d69d11055b513604b1d1

Former-commit-id: 1c058a68e9d2831ac4c607916dca3494b9fbd911