Fix conditional for glClear call
Previously the call to glClear was executed when there was no
available space (i.e. space which the Bitmap would *not* be drawn in
on the background of the launcher), rather than when there was.
Inverting the checks fixes this problem.

Also, remove unnecessary Handler since the call to updateWallpapers
is done via one-way binder interface and wrapped in a synchronize
block anyways. Putting it on the Handler just put it on the main
looper of the context that created WallpaperManager, which isn't
necessary.

Change-Id: Ic7a323303ec6e354d1ef245eec3434ff7128432d

Former-commit-id: 6762a441ef9c764f3bfee4201742e80aa6621b89