Fix memory leak and slowness in keyguard
This fixes an issue where the mUpdateSeekBars runnable was being
repeatedly posted to the handler while keeping a reference to
the transport control (and hence keyguard).  Even after the view
was detached from keyguard it would continue to post events that
would ultimately keep the instance of keyguard from being GC'd and
cause keyguard to slow down.

Fixes 11169793

Change-Id: If2bb238adf6d78101589926b60a61d9f187c6d1c

Former-commit-id: b2c64690a3d1fe4d2100b34e9f3d823985349eed