Add system partition check for vpn dialogs
The VPN permission check required internal knowledge that
other checks in the system would ensure that
package "com.android.vpndialogs" was genuine. In the
off chance those other checks change, or someone is
able to spoof the package name, this will at least
check to see that the app is on the system image; one
more hurdle to jump.

The original code for this change came from
https://android-review.googlesource.com/62270

Change-Id: I55580bee0b30036b0fee9ca4e43de9b736b194fe
Signed-off-by: William Roberts <wroberts@tresys.com>

Former-commit-id: 212a195f008688c5483bf369d5a96753d582a24b