TTS API: Replace "voice id" with "voice name"
"Voice id" and "voice name" are both used to reference a voice in a
speech synthesis request. Voice id was a random integer, where
voice name is human readable string, that provides more debug
information and readability. Also, it's expected that voice name
will stay consistent, and won't change during the life of the speech
connection. Though, it may disappear.

Change-Id: I180296d413a18301cead1c8e3212de2bd0c7e32d

Former-commit-id: 97cd647fcd89c1a7a4448c4fc4d9ad906f8cf2a1