Share pack historical sorting using wrong keys.
The ActivityChooserModel keeps a history of the last fifty
share targets and based on past usage orders the targets in
the UI. The soring implementation is using a map for improving
performance. However, the activities in this map were keyed
on the package name but there maybe more that one share
target per package. Thus, the sorting was generating bad
results. Now the unique component name is used.

bug:11195578

Change-Id: I8c7018fea168b7253ddbe57b477028368726e75e

Former-commit-id: 27f592df8ce604c62327cc93ee5adac3152a0bd2