Before requerying a cursor, check if it is closed.
Before updating a cursor,
it needs to be checked if it is valid.
Currently a defence code only checks
if it is not null.
But in case of "Cursor", it is not valid,
if it is closed.
So just like in CursorAdapter.onContentChanged(),
add the additional defence code to check validation.

Change-Id: I1302f1b6d009897815d4bdb2f20f625688cf9901

Former-commit-id: 912ab78c5fa65dcdebaf5d4b26ff4964b6785907