PreferencesFragment showing the breadcrumb area in single pane mode.
If the PreferenceActivity is running in a single pane mode we are
not showing the headers and the breadcrumb area. However, when this
activity is restarted and has a saved state to restore we are trying
to use headers even in a single pane mode. As a result the breadcrumb
area is shown and the content is shifted to the bottom with an empty
space at the top. This change ignores the saved headers from the
saved instance state in a single pane mode. Note that in such a case
these headers are null anyway as we do not use them.

bug:11242762

Change-Id: I2828bc82762695d9c93fb4ca43933598a9b12b87

Former-commit-id: 2b8795a9db2d494c38d34fd113a9153880a555c0