Print service meta-data changes not handled properly.
The package monitor in the PrintManagerService was not overriding
the correct callback to capture changes in the print service package
including meta-data. Hence, there was a missed signal the the old
meta-data was used.

The implication is that if a vendor adds new feature such as custom
print options activity or print service settings that were not there
before, then they will not be show until the device reboots.

bug:11772977

Change-Id: Idf756c430f30ae6d16760a257c199a72a52cf5c5

Former-commit-id: 8ef0c14858de7a178aac2a5a38fe1651d3592d7f