Make sure to turn off led after pulse()
setLightLocked() will update mColor. This will cause the led to not be
turned off after the pulse.
When notificationManager starts a attention pulse the led will be kept
on. Since attention is of highest prio nothing can turn it off.
This is fixed by resetting mColor back to 0 after the call.

Change-Id: Id60ef96e5c21b47c23002f0bcf2fae7fb3f2ca10

Former-commit-id: e8467194dc8d56b5cf418b025b041737754b98b3