ProgressBar is important for accessibility by default.
If a view does not have callbacks, or content description, or does not draw
content, and it is marked as important for accessibility auto, then the system
decides it is not important and does not report it. Apparently progress bar
draws content that means something and it should be important for accessibility
by default.

Change-Id: Icd3837fb8b9e208c98b90707f3b195622d71949e
(cherry picked from commit 98626eaa8301aaa8e366174d07001ddda20aeeaf [formerly 7face75f2c655a371a4dcdf6063886f92c637ce4])

Former-commit-id: 6a1846a084fd09b3ade3fa80a49c29dcbe7cd60d